+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @openpkg-ts/react
 
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @openpkg-ts/sdk@0.33.0
+  - @openpkg-ts/spec@0.33.0
+
 ## 0.2.4
 
 ### Patch Changes